Create grab ids by year script
It turned out that getting sample arXiv papers by years is quite a challenge.
The arxiv API doesn't have a 'year' parameter so we had to resort to
different methods. We took the 'gr-qc' (general relativity and quantum
cosmology) category and pretty much went through all the papers in the
category. For each year that the papers were published, the first 50 returned
papers are saved.